Create force graph component
