+++ web/src/components/ForceGraph/ForceGraph.stories.tsx
@@ -1,0 +1,7 @@
+import ForceGraph from './ForceGraph'
+
+export const generated = () => {
+  return <ForceGraph />
+}
+
+export default { title: 'Components/ForceGraph' }
